+++ plusplusbot/commands/command.py
