+++ plusplusbot/commands/general_commands/__init__.py
@@ -1,0 +1,6 @@
+from .help_command import HelpCommand
+
+
+general_commands_list = [
+    HelpCommand,
+]
